Move Dataset and MinMaxScaler classes to separate files
