+++ pyLOM/NN/__init__.py
@@ -27,7 +27,9 @@ PIN_MEMORY = True if torch.cuda.is_available() else False
 from .                                import GAVI
 
 from .pipeline                        import Pipeline, ClusteredPipeline
-from .utils                           import Dataset, MinMaxScaler, select_device, betaLinearScheduler, create_results_folder, set_seed
+from .dataset                         import Dataset
+from .scalers                         import MinMaxScaler
+from .utils                           import select_device, betaLinearScheduler, create_results_folder, set_seed
 from .utils                           import tanh, relu, elu, sigmoid, leakyRelu, silu
 
 from .optimizer                       import OptunaOptimizer, OptunaStudyManager
